refactor(hf_model_adapter): remove duplicate code
- loading a model with registry is now in models/utils.py
- both tests and hf_adapter call util method
