+++ tests/checkpointing/test_checkpoint_conversion.py
@@ -12,9 +12,11 @@ from modalities.config.component_factory import ComponentFactory
 from modalities.config.config import load_app_config_dict
 from modalities.config.pydanctic_if_types import PydanticPytorchModuleType
 from modalities.models.huggingface_adapters.hf_adapter import HFAdapter, HFAdapterConfig
+from modalities.models.utils import get_model_from_config
 from modalities.registry.components import COMPONENTS
 from modalities.registry.registry import Registry
 from tests.conftest import _ROOT_DIR
+
 
 @pytest.fixture()
 def set_env():
@@ -52,14 +54,8 @@ def config_dict(config_file_path):
 
 
 @pytest.fixture()
-def initialized_model(set_env, component_factory, config_dict):
-    class ModelConfig(BaseModel):
-        model: PydanticPytorchModuleType
-
-    components = component_factory.build_components(
-        config_dict=config_dict, components_model_type=ModelConfig
-    )
-    return components.model
+def initialized_model(set_env, config_dict):
+    return get_model_from_config(config=config_dict)
 
 
 @pytest.fixture()
@@ -72,8 +68,7 @@ def checkpoint_conversion(tmp_path, initialized_model, config_file_path):
         config_file_path=config_file_path,
         output_hf_checkpoint_dir=output_hf_checkpoint_dir,
     )
-    checkpoint_conversion.config_dict["checkpointed_model"]["config"][
-        "checkpoint_path"] = model_file_path
+    checkpoint_conversion.config_dict["checkpointed_model"]["config"]["checkpoint_path"] = model_file_path
     return checkpoint_conversion
 
 
@@ -91,8 +86,9 @@ def hf_model(checkpoint_conversion):
 def hf_model_from_checkpoint(checkpoint_conversion, pytorch_model, device):
     AutoConfig.register("modalities", HFAdapterConfig)
     AutoModelForCausalLM.register(HFAdapterConfig, HFAdapter)
-    hf_model_from_checkpoint = AutoModelForCausalLM.from_pretrained(checkpoint_conversion.output_hf_checkpoint_dir,
-                                                                    torch_dtype=pytorch_model.lm_head.weight.dtype)
+    hf_model_from_checkpoint = AutoModelForCausalLM.from_pretrained(
+        checkpoint_conversion.output_hf_checkpoint_dir, torch_dtype=pytorch_model.lm_head.weight.dtype
+    )
     hf_model_from_checkpoint = hf_model_from_checkpoint.to(device)
     return hf_model_from_checkpoint
 
@@ -110,11 +106,11 @@ def test_hf_and_pytorch_models_are_the_same_after_init(hf_model, pytorch_model, 
 
 
 def test_models_before_and_after_conversion_produce_same_output(
-        device,
-        pytorch_model,
-        hf_model,
-        hf_model_from_checkpoint,
-        test_tensor,
+    device,
+    pytorch_model,
+    hf_model,
+    hf_model_from_checkpoint,
+    test_tensor,
 ):
     pytorch_model = put_model_to_eval_mode(pytorch_model, device)
     hf_model = put_model_to_eval_mode(hf_model, device)
